feat: 1450 http client refactoring
